correction commissaire
la vente tourne bien mais dans le vide pour l'instant

+++ server/defs.h
@@ -15,7 +15,7 @@ int objInSale;
 OBJET objs[TMAX];
 pthread_mutex_t mutexObjs;
 
-int bid;
+float bid;
 pthread_cond_t condBid;
 pthread_mutex_t mutexBid;
